new levels but bug w updatables :(
added levels in between existing ones - difficulty scale too hard
when a timer calls a function that creates more updatables, i think that causes some issues
i tried creating a queue list for updatables to be added onto next frame, but idt that worked

+++ src/gameplay/gameObjects/puzzlePiece/Side.java
@@ -110,7 +110,8 @@ public class Side {
     public void disconnect() { 
         connected = false;
         tweenPercent = 0;
-        connectTween.delete();
+        if (connectTween != null)
+            connectTween.delete();
 
     }
     public void connect(boolean playAnimation) {
